+++ src/key-point/vpBasicKeyPoint.h
@@ -184,14 +184,14 @@ class VISP_EXPORT vpBasicKeyPoint
 
      \return the number of reference points.
     */
-    inline unsigned int getReferencePointNumber() const {return referenceImagePointsList.size();};
+    inline unsigned int getReferencePointNumber() const {return (unsigned int)referenceImagePointsList.size();};
 
     /*!
      Get the number of matched points.
 
      \return the number of matched points.
     */
-    inline unsigned int getMatchedPointNumber() const {return matchedReferencePoints.size();};
+    inline unsigned int getMatchedPointNumber() const {return (unsigned int)matchedReferencePoints.size();};
 
     /*!
       Return the vector of reference image point.
